fixing vcpkg environment conflicts. will re-added prior committed coded

+++ core/vm/arch/memory.h
@@ -220,7 +220,7 @@ class MemoryManager {
   }
   
   static size_t* AllocateObject(const long obj_id, size_t* op_stack, long stack_pos, bool collect = true);
-  static size_t* AllocateArray(const long size, const MemoryType type, size_t* op_stack, long stack_pos, bool collect = true);
+  static size_t* AllocateArray(const size_t size, const MemoryType type, size_t* op_stack, long stack_pos, bool collect = true);
   
   // object verification
   static size_t* ValidObjectCast(size_t* mem, long to_id, long* cls_hierarchy, long** cls_interfaces);
